Fanotify, Inotify: Send remote request when adding a local watch
Add the capability in fanotify and inotify to send a request to server to
add a remote watch when a local watch is added.

Specifically, add the "fsnotify_update" function within the
"inotify_add_watch" (for inotify) and "fanotify_mark" (for fanotify) system
calls, after a local watch has been added or removed successfully.

The "fsnotify_update" function always sends the aggregated watch for an
inode that is monitored for an event.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>